requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
idna==3.6
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.5
Pygments==2.17.2
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
spotipy==2.23.0
textual==0.49.0
textual-dev==1.4.0
typing_extensions==4.9.0
uc-micro-py==1.0.2
urllib3==2.2.0
yarl==1.9.4
